+++ src/main/java/io/github/oliviercailloux/y2018/apartments/gui/AskOpinionForUtility.java
@@ -120,7 +120,7 @@ public class AskOpinionForUtility {
 		
 		LOGGER.info("Begining the Layout.");
 		
-		Layout2 lay = new Layout2(avf);
+		LayoutApartmentGUI lay = new LayoutApartmentGUI(avf);
 		lay.displayAppart();
 
 	}
